+++ linuxdoom-1.10/r_segs.h
@@ -32,8 +32,8 @@
 void
 R_RenderMaskedSegRange
 ( drawseg_t*	ds,
-  int		x1,
-  int		x2 );
+  short		x1,
+  short		x2 );
 
 
 #endif
